Extract parts of mormalize methods, to make it clearer they're different
This is a baffling piece of design in the original library. These aren't
all same-name methods in the original, they're parts of other functions.
However, they're chunks of code that would be thought to do the same
thing: normalize a given data frame / dependencies pair. In fact, two of
them only do the first non-trivial step of a normalization, then stop.
This is silly, and something I'm looking to delete.