Speed up relationship-lossless concatenation test for database schemas
Using lapply, vapply etc. over a list of schemas is slow, because for
each element it then needs to subset ("[[") to the one schema; under
current subsetting methods this constructs a new schema object, does
validation, etc. If we work with the attrs and keys instead, this
doesn't occur, so we skip a lot of redundant validation.