Add attrs_class attribute to function_dependency and schema objects
We want to split decomposition into two steps: creation/instantiation,
giving a relation/database object from the given schema, and insertion,
which adds the data. We'd thus like decomposition to be equivalent to
creation, followed by insertion.

However, we also want the property that inserting zero records has no
effect. Since a data frame passed into discover etc. can have zero
records, this means that creation must be able to pass on default
classes for the attributes. In turn, because of how the main autodb
workflow works, this means that all of the objects mentioned above must
be able to carry over the class information from initial data frames.

Since the attribute class information also give the attribute order, the
attrs_order attribute is now technically redundant in these classes.
However, in the case where the user is creating one of these objects
manually, requiring them to provide class information with no default is
unreasonable, so we keep attrs_order as an input parameter for user
convenience. We can remove it from the actual objects, however, to avoid
the data redundancy, so attrs_order and `attrs_order<-` methods now
refer to the attrs_class attribute instead.