Remove attribute class information from FD/schema classes
Default attribute class information turned out to be rather ungainly:
for example, it doesn't play nicely with the default behaviour of rbind.
Having class information contained in functional dependencies is also a
little silly. The plan is to get rid of the need for class information
by replacing the "insertion of empty data does nothing" constraint with
an "insertion of empty data into a non-empty relations does nothing"
constraint, since this removes the issue when inserting a zero-record
original table.