Add records and records<- for relations, remove relation/database interior fetching in tests
"Interior fetching" here refers to going into list elements, and making
use of the knowledge that relations are stored as df-keys pairs to
access data and key information. We're about to make this impossible by
introducing subsetting methods for relations and databases, and it's
simpler to remove these from the tests beforehand.