Make DiagrammeR optional in examples and vignette
Packages in "Suggests" are supposed to be optional, even for examples,
vignettes, tests, etc. We therefore wrap DiagrammeR calls in
requireNamespace, as suggested in the documentation for writing R
extensions.

hedgehog is also in "Suggests", but making it optional would leave the
test suite rather bare, so I'm leaving it as it is.