Remove data class requirement constraint for foreign keys
Foreign key check previously checked that referring and referred
attributes have the same class. Since we're using an dynamically-typed
langauge (R), it makes more sense to keep to the usual practice for the
language, and allow different data classes to reference each other. This
does introduce the issue that key checks rely on merge.data.frame, so
for classes that don't work properly in merges, autodb breaks. I think
saying "let the user beware" is the appropriate decision here, since the
issue is in the base R language rather than autodb itself. It's just
unfortunate that some classes in the base language don't work properly
in merges (see changes to the vignette).