Make local copies of used snakecase functionality
This consists of snakecase functions, simplified to just the code used
to call to_snake_case. This removes the snakecase dependency with a
dependency on just stringr. The planned next step is to replace this
with just dependencies on stringi etc., until we have a much-simplified
version of to_snake_case, that we can replace with a simple function
that only does the conversions needed for Graphviz labels.