Add error message for relation attributes not in attrs_order, sort keys and attributes properly
I'm not sure why I wasn't sorting keys and attributes before, let alone
why I was instead throwing an error if the user hadn't pre-sorted the
attributes accordingly. In any case, sorting brings the system closer to
the intended design, where ordering is something the user shouldn't have
to worry about, because it's only used to simplify print/plot displays:
the user should be able to think of attributes and keys as unordered
sets.