Make subset reassignment consistent for relation schemas
In other words, subset reassignment returns another relation schema,
with only the contents of the subset changed. I think this is the last
of the default generic functions that can break class containment
without writing methods for them, not counting functions like unclass,
which are explicitly intended to break it.