Add data frame functions that handle zero columns correctly
duplicated(), and functions that use it, currently work incorrectly for
data frames with zero columns, because it always returns an empty
vector, regardless of the number of rows. This has been reported as a
bug, and fixed in trunk, so potentially will be fixed in R 4.2. For now,
we need these fixed versions, and might keep them afterwards for
compatibility with older versions of R.