Remove pairwise check when merging attrs_order sets
This didn't work anyway, because as1 and as2 used the pair values,
instead of using them to index into ordered_sets. The cases it checks
for, where the attrs_order sets can't be merged pairwise, get detected
anyway in the main part of the process, and doing only the latter means
they at least get merged as much as they can at the beginning. The pair
step is, I think, a leftover from before I added the main part.