Remove name attribute from database class, check gv name argument
The name attribute was pretty tacked on; it was only for naming a graph
when plotting, anticipating a future feature where graphs can be merged
together. That's a while off, and since the name is only used for
plotting it's better off in the plotting function anyway, where it's
already present for other class methods.