Let powerset have limit on set size, reduce with it in mind
Currently this isn't used, and using a size-restricted power in discover
breaks everything, because powerset navigation assumes no set size
limit, so everything can be done in terms of bitsets. Allowing limit
calls for a more general navigation approach, possibly based on match(),
which is slow, so I'm hoping for another way to do it. In any case, the
plan is to move powerset-related code into its own file/module first.