Simplify cross_intersection
cross_intersection is a likely source of errors when trying to allow a
maximum detset size, because it constructs seed nodes mostly
independently from the powerset. The changes here are simple incremental
changes that don't break anything. The name changes making it clear
that a bitset index is currently used as a node; converting to a node
breaks current behaviour. Working out why is probably the key to getting
maximum detset sizes to work.