Speed up powerset construction
As ever, using match when said matches are known to be in order results
in bad scaling. In this case, we can remove the need for a match by
making use of the bitset_index information, which we calculate anyway,
and is a simpler match. If this one becomes a problem, we can consider
writing some form of "ordered match", where we know that successful
matches will be in order.