Change powerset reduction WRT what bitset_index represents
Previously, bitset_index represented what each node in the original,
non-detset-size-limited powerset would map to in the current graph, if
anything. Now, it represents what each node in the current-cardinality,
non-detset-size-limited powerset maps to in the current graph, so it no
longer accounts for the original cardinality before reductions.
Cardinality tracking is already handled by converting attribute index
conversion, so we only need to track size-limiting to allow handling of
det-size limiting.