Rewrite detset limit code for discover to never slow DFD down
The original implementation worked by removing search nodes for sets
above the given limit. For attributes with sets larger than that, or no
sets at all, this caused DFD to "ping-pong" again the limit boundary,
which is much less efficient than searching above the bounday as usual.

The new implementation works by forbidding sets above the given size
from being search path seeds, so that nodes for sets above the limit are
only visited when it helps resolves sets within the limit. This results
in having to filter out that the above-limit dependencies at the end of
the full search, and rarely results in the search being faster than not
giving a limit, but it never results in the search taking longer.