Remove addition seed pruning in no-max-nondeps case
Filtering to only seeds with non-strictly-negate category was added by
the Python autonormalize library: it's not present in the original
Metanome code. This is not surprising, because in that case it's
impossible for any nodes to have negative category: at seed generation
time, no nodes can be left as candidates, so if there are no maximal
non-dependencies, there can't be any known non-dependencies at all. In
fact, given how seeds are determined, the only node knowledge possible
is of single-attribute nodes being minimal depednencies.