Re-introduce initial check for simple keys, skip in skip_bijections
I can't remember why I remove the simple key check originally, maybe
because I thought it didn't do much. Of course, if we can skip entire
dependants/determinants, it can do a lot. Here we just skip all but the
first simple key as both dependants and determinants. Not added yet is
that we don't need to use the first key as a determinant either: we know
that it must make a minimal FD for all non-fixed dependants, so we can
begin each LHS search with some node categories known, whether we choose
to skip bijections or not.