Start with partial difference information in tree search, using restarts
The idea is that new difference information can be found and corporated
during the tree search, with no need for backtracking to visited nodes
to update them. I've found that this isn't the case, so for now I'm
taking the simpler option of restarting when new information is found.
This is not a quick solution -- it's slower than DFD, for example -- but
it works.