Move generic partition functions to new module
partition_rank is used by both DFD and FDHits, so can easily be
considered as "generic". stripped_partition_product is currently only
used by DFD, but the intention for moving FDHits() into discover() at
this point was to more easily allow it to make use of existing partition
combination code.