Fix FDHits partition node names
Using a full string, like in this fix, can be slow, which is why I
replaced it with integer names in DFD. The disadvantage there was that
it limits the search to 31 candidate determinant attributes. The
bit-packing approach I'm replacing here was intended to allow for any
number of attributes up to integer.max, but multibyte characters, and
larger numbers being converted to up to four characters each -- so that
the conversion isn't invertible -- make this infeasible. I'm changing it
back to full strings for now.