Add a trace cache that's built and not used
This is a first step in substantially reducing the handler's caching.

FDHits has a nice property. If we began with all of the difference sets,
we could show the path the search will take, as it is a depth-first tree
search. Starting with only a sample means that we begin with some
partial tree, possibly one that isn't a subset of the full tree. When we
add more difference sets, the effect is to grow new child trees from
existing nodes, and the clever part of FDHits is to order the search
so that these never appear on nodes we've already visited. Therefore, at
the end of the search, we again have a search trace that is a
depth-first search on a revealed tree, with no backtracking.

This DFS-like structure means that we can make cache simplifications
that we could do for DFS: a node's information is only needed for
visiting itself and its children. We can therefore make our cache
structured around the current trace, letting us discard information when
no longer needed, and greatly reducing the memory usage, and the amount
of new information to add when we add more difference sets.

This is inspired by the Metanome Rust code for FDHits, which takes a
similar approach.