Prepare trace critical cache using predecessor
Due to tree-like traversal, the predecessor is the node's parent, so is
similar to the current node's information, and getting from one to the
other is cheap compared to calculating from scratch. This gives a large
speedup on critical edge calculations, which previously dominated the
runtime for small data sets.