Skip match.arg step for packBits
At the time of commit, packBits take up an eighth of the runtime of
`discover(nudge)` when using FDHits. The vast majority of that is spent in
a call to `match.arg`, and the vast majority of that is handling the fact
that the `choices` argument is missing. Using the `.Internal` call
directionly is nasty, but it skips this step entirely.