Convert vignettes to Quarto, use dot chunks instead of DiagrammeR
Using DiagrammeR to call GraphViz became an issue when splitting the
vignette up into several, because the Javascript required for HTML
widgets takes up about 2 MB per HTML file, which put the package size
well above the CRAN limit. Switching to dot code chunks removes this
overhead.

The README is currently still generated from an RMarkdown file using
DiagrammeR, because the tooling for generating the README from Quarto is
not as straightforward.