Revert to using DiagrammeR in vignettes, keep quarto
As of writing, quarto is a pain for the use case I have here: vignettes
where we generate GraphViz code, then render it within a dots code
block. This is because this involves saving the code in a file, which is
made difficult by pkgdown building in a temporary directory. We could
get around this by rendering the images ahead of time, e.g. as a script
in data-raw, but this involves choosing dimensions for the resulting
image/snapshot. This is fiddly work, since making the image too narrow
results in table names etc. overflowing their assigned area. I'd rather
avoid that, so for now I'm going back to using DiagrammeR, and putting
up with the warning about the vignette folder's size.