Simplify and rearrange discover() tests
Some tests made others redundant, I've rearranged the rest into rough
thematic groups, I finally added a function for the common pattern of
running discovery with a timeout, and I move another test into the large
general metamorphic test. Generally, I want to move more of the tests
into this large one, or an equivalent one for accuracy < 1. This looks
awful to someone used to TDD, since it looks like we're conflating
together tests that should be independent. However, that's actually not
the case: the individual tests are separate, we're just giving them the
same (generated) inputs. This means we're not wasting computing time by
generating inputs for each of them separately.